// frontend/src/api/generated/DocumentVersionDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DocumentVersionDto = { id: string, document_id: string, number: bigint, file_id: string, note: string, uploaded_by: string, 
/**
 * Display name of the uploader.
 */
uploaded_by_name: string, uploaded_at: string, scan_status: string, size: bigint, mime: string, };